requirements.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.7
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.1
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
cryptography==37.0.4
decorator==5.1.1
distlib==0.3.1
Django==4.0.6
django-cors-headers==3.13.0
django-js-asset==1.2.2
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.9.1
filelock==3.0.12
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
gunicorn==20.1.0
h11==0.13.0
hubspot-api-client==5.1.0
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
oauthlib==3.2.0
outcome==1.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.13.0
python-http-client==3.3.7
pytz==2020.1
qrcode==7.3.1
razorpay==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
sendgrid==6.9.7
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.17
sqlparse==0.3.1
stack-data==0.3.0
starkbank-ecdsa==2.0.3
termcolor==1.1.0
toml==0.10.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==0.15.4
whitenoise==5.1.0
wrapt==1.12.1
wsproto==1.2.0
